[refactor]: DropDownIcon을 common 폴더로 이동
